Move everything in custom hook
